refactoring with an attempt to follow good practices and standards :)

+++ src/main/java/leonardo/labutilities/qualitylabpro/services/AnalyticsService.java
@@ -1,20 +1,15 @@
 package leonardo.labutilities.qualitylabpro.services;
 
-import jakarta.annotation.PostConstruct;
 import leonardo.labutilities.qualitylabpro.analytics.Analytics;
+import leonardo.labutilities.qualitylabpro.infra.exception.ErrorHandling;
 import leonardo.labutilities.qualitylabpro.records.valuesOf.ValuesOfLevels;
 import leonardo.labutilities.qualitylabpro.records.valuesOf.ValuesOfLevelsList;
 import leonardo.labutilities.qualitylabpro.repositories.AnalyticRepository;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
-import org.springframework.web.bind.annotation.PathVariable;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Objects;
-
-import static leonardo.labutilities.qualitylabpro.analytics.Analytics.analyticsHashMap;
 
 @RequiredArgsConstructor
 @Service
@@ -25,6 +20,7 @@ public class AnalyticsService {
     public Analytics sendValues(ValuesOfLevels values) {
         return analyticRepository.save(new Analytics(values));
     }
+
     public List<Analytics> sendValuesList(List<ValuesOfLevels> valuesOfLevelsList) {
         List<Analytics> analyticsList = new ArrayList<>();
         for (ValuesOfLevels value : valuesOfLevelsList) {
@@ -34,36 +30,40 @@ public class AnalyticsService {
         }
         return analyticsList;
     }
-    public void deleteValues(Long id){
+
+    public void deleteValues(Long id) {
+        if (!analyticRepository.existsById(id)) {
+            throw new RuntimeException("invalid values or not exists");
+        }
         analyticRepository.deleteById(id);
     }
-    public void deleteValuesAll(){
+
+    public void deleteValuesAll() {
         analyticRepository.deleteAll();
     }
 
     public List<ValuesOfLevelsList> getResultsAll() {
-        return analyticRepository.findAll().stream().map(ValuesOfLevelsList::new).toList();
+        if (!analyticRepository.findAll().isEmpty()) {
+            return analyticRepository.findAll().stream().map(ValuesOfLevelsList::new).toList();
+        }
+        throw new ErrorHandling.NoContentException("List is empty");
     }
 
     public List<ValuesOfLevelsList> getResultsByName(String name) {
-        if (analyticRepository.existsByName(name)) {
-            return analyticRepository.findAll().stream()
-                    .filter(s -> Objects.equals(s.getName(), name.toUpperCase())).map(ValuesOfLevelsList::new).toList();
+        if (!analyticRepository.existsByName(name)) {
+            throw new RuntimeException("Test name not exist in Database");
         }
-        throw new RuntimeException("Test name not exist in Database");
+        return analyticRepository.findAllByName(name).stream()
+                .map(ValuesOfLevelsList::new).toList();
     }
 
     public String getResultsByNameLevel1(String name) {
         if (analyticRepository.existsByName(name)) {
-            var list = analyticRepository.findAll().stream().map(ValuesOfLevelsList::new).toList();
-            List<String> analyticListByNameList = new ArrayList<>();
-            for (ValuesOfLevelsList item : list) {
-                if (item.name().contains(name.toUpperCase())) {
-                    analyticListByNameList.add(item.date() + " Nível1: " + item.name() + " ---> " + item.normalValue() + " : " +
-                            item.normalValid() + ", Regra: " + item.normalObs());
-                }
-            }
-            return analyticListByNameList.toString();
+            var listByNameLevel1 = analyticRepository.findAllByName(name).stream().map(ValuesOfLevelsList::new).toList();
+            return listByNameLevel1.stream()
+                    .map(analytic -> analytic.date() + "Level1: " + analytic.name() + " ---> "
+                            + analytic.normalValue() + " : " + analytic.normalValid() + ", Regra: "
+                            + analytic.normalObs()).toList().toString();
         }
         throw new RuntimeException("Test name not exist in Database");
     }
@@ -74,7 +74,8 @@ public class AnalyticsService {
             List<String> analyticListByNameList = new ArrayList<>();
             for (ValuesOfLevelsList item : list) {
                 if (item.name().contains(name.toUpperCase())) {
-                    analyticListByNameList.add(item.date() + " Nível2: " + item.name() + " ---> " + item.highValue() + " : " +
+                    analyticListByNameList.add(item.date() + " Nível2: " + item.name() + " ---> "
+                            + item.highValue() + " : " +
                             item.highValid() + ", Regra: " + item.highObs());
                 }
             }
